update docs, move some files
