+++ .npmignore
@@ -1,7 +1,18 @@
+# Logs
+*.log
+
+# Development files
 src
 examples
+
+# Test
 coverage
-.babelrc
+test
+
+# Config
 .gitignore
+.babelrc
 webpack.config.js
-.DS_Store
+.circleci
+jest.setup.js
+.eslintrc
