+++ test/mocks/MockSynthesis.js
